+++ Cargo.toml
@@ -177,6 +177,15 @@ semver = "1.0"
 serde_json = { version = "1", optional = true }
 sha-1 = { version = "0.10", optional = true }
 socket2 = "0.5"
+tracing = { version = "0.1", optional = true }
+tracing-futures = { version = "0.2", optional = true }
+url = "2.4"
+urlencoding = "2.1"
+# DNS Features
+trust-dns-resolver = { version = "0.23", optional = true, features = ["tokio"] }
+hickory-resolver = { version = "0.24.1", optional = true, features = ["tokio"] }
+fred-macros = "0.1"
+# Tokio  Dependencies
 tokio = { version = "1.34", features = [
     "net",
     "sync",
@@ -188,18 +197,18 @@ tokio-native-tls = { version = "0.3", optional = true }
 tokio-rustls = { version = "0.26", optional = true, default-features = false }
 tokio-stream = "0.1"
 tokio-util = { version = "0.7", features = ["codec"] }
-tracing = { version = "0.1", optional = true }
-tracing-futures = { version = "0.2", optional = true }
-trust-dns-resolver = { version = "0.23", optional = true, features = ["tokio"] }
-hickory-resolver = { version = "0.24.1", optional = true, features = ["tokio"] }
-url = "2.4"
-urlencoding = "2.1"
-fred-macros = "0.1"
+# Glommio Dependencies
 glommio = { version = "0.9.0", optional = true }
 futures-io = { version = "0.3", optional = true }
 pin-project = { version = "1.1.5", optional = true }
 oneshot = { version = "0.1.8", optional = true, features = ["async"] }
 futures-lite = { version = "2.3", optional = true }
+# Monoio Dependencies
+monoio = { version = "0.2.4", optional = true, features = ["bytes"] }
+monoio-codec = { version = "0.3.4", optional = true }
+monoio-native-tls = { version = "0.4.0", optional = true }
+monoio-rustls = { version = "0.4.0", optional = true }
+# Smol Dependencies
 
 [dev-dependencies]
 axum = { version = "0.7", features = ["macros"] }
