+++ src/runtime/glommio/interfaces.rs
@@ -4,13 +4,12 @@ use crate::{
   clients::WithOptions,
   commands,
   error::RedisError,
-  glommio::compat::spawn_into,
   interfaces::{RedisResult, Resp3Frame},
   modules::inner::RedisClientInner,
   prelude::default_send_command,
   protocol::command::RedisCommand,
   router::commands as router_commands,
-  runtime::{spawn, BroadcastReceiver, JoinHandle, RefCount},
+  runtime::{glommio::compat::spawn_into, spawn, BroadcastReceiver, JoinHandle, RefCount},
   types::{
     ClientState,
     ConnectHandle,
